pyproject.toml
[tool.poetry]
name = "betterglauco"
version = "0.1.0"
description = ""
authors = ["divirjo <worldoftyto@gmail.com>"]
readme = "README.md"
packages = [{include = "betterglauco"}]

[tool.poetry.dependencies]
python = "3.12.*"
django = "^4.2"
django-crispy-forms = "^2.0"
crispy-tailwind = "^0.5.0"
pytz = "^2023.3"
requests = "^2.28.2"
babel = "^2.12.1"
django-tables2 = "^2.5.3"
psycopg2-binary = "^2.9.6"
protobuf = "^4.22.3"
python-dotenv = "^1.0.0"
django-filter = "^23.5"
pandas = "^2.1.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.0"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations','manage.py']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
migrate = 'python manage.py makemigrations && python manage.py migrate'
run = 'python manage.py runserver'
pre_test = 'task lint'
test = 'pytest -s -x --cov=betterGlauco -vv'
post_test = 'coverage html'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
virtualenvs.in-project = true
